Update Unit Tests and Line Control
Update Line Control to also keep track of the data type of the channel which will be needed for users to be able to select timestamp type as it is not a tagbus type.

Also added lvtests for test VIs that didn't have them.

Deleted Find Duplicate Test.vi as it was really just testing VIs in tag editor core and not the module.